pyproject.toml
[project]
name = "ptt"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "arkerphyo92", email = "arkerphyo92@gmail.com"},
]
dependencies = [
    "django",
    "django-extensions",
    "pillow>=11.1.0",
    "django-silk>=5.3.2",
    "django-jazzmin>=3.0.1",
    "django-modeltranslation>=0.19.12",
    "django-ninja>=1.3.0",
    "django-ninja-jwt>=5.3.5",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
